updated notebooks
